+++ dmarc-report/dmarc-report.d/local/share/cook/bin/configure-parsedmarc.sh
@@ -50,7 +50,9 @@ sep=$'\001'
   sed "s${sep}%%imapuser%%${sep}$IMAPUSER${sep}g" | \
   sed "s${sep}%%imappass%%${sep}$IMAPPASS${sep}g" | \
   sed "s${sep}%%imapfolder%%${sep}$IMAPFOLDER${sep}g" | \
-  sed "s${sep}%%outputfolder%%${sep}$OUTPUTFOLDER${sep}g" \
+  sed "s${sep}%%outputfolder%%${sep}$OUTPUTFOLDER${sep}g" | \
+  sed "s${sep}%%ip%%${sep}$IP${sep}g" | \
+  sed "s${sep}%%zincport%%${sep}$ZINCPORT${sep}g" \
   > /usr/local/etc/parsedmarc.ini
 
 # set the ownership of the config file. This could be changed